requirements.txt
# ================================
# 闲鱼自动回复系统 - Python依赖包
# ================================

# ==================== 核心Web框架 ====================
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# ==================== 日志记录 ====================
loguru>=0.7.0

# ==================== 网络通信 ====================
websockets>=10.0,<13.0
aiohttp>=3.9.0
requests>=2.31.0
httpx>=0.25.0

# ==================== 配置文件处理 ====================
PyYAML>=6.0.0

# ==================== JavaScript执行引擎 ====================
PyExecJS>=1.5.1

# ==================== 协议缓冲区解析 ====================
blackboxprotobuf>=1.0.1

# ==================== 系统监控 ====================
psutil>=5.9.0

# ==================== 文件上传支持 ====================
python-multipart>=0.0.6

# ==================== AI回复引擎 ====================
openai>=1.50.0

# ==================== 图像处理 ====================
Pillow>=10.0.0
qrcode[pil]>=7.4.2

# ==================== 浏览器自动化 ====================
playwright>=1.40.0
DrissionPage>=4.0.0

# ==================== 重试机制 ====================
tenacity>=8.2.0

# ==================== 加密和安全 ====================
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0

# ==================== 时间处理 ====================
python-dateutil>=2.8.2

# ==================== Excel文件处理 ====================
pandas>=2.0.0
openpyxl>=3.1.0

# ==================== 邮件发送 ====================
email-validator>=2.0.0

# ==================== 数据处理和验证 ====================
xlsxwriter>=3.1.0

# ==================== 构建二进制扩展模块（可选） ====================
# 用于编译性能关键模块，提升运行效率
# 如果不需要编译功能，可以注释掉以下依赖
nuitka>=2.7
ordered-set>=4.1.0
zstandard>=0.22.0

# ==================== 版本说明 ====================
# 本文件包含闲鱼自动回复系统的所有必需依赖
# 版本号已经过测试验证，确保兼容性和稳定性
#
# Python版本要求: 3.11+
#
# 安装命令:
# pip install --upgrade pip
# pip install -r requirements.txt
# playwright install chromium
# playwright install-deps chromium  # Linux系统需要
#
# 注意：DrissionPage需要Chrome/Chromium浏览器支持滑块验证功能
# 如果系统没有Chrome，请先安装Chrome浏览器
#
# ==================== 内置模块说明 ====================
# 以下模块是Python内置模块，无需安装：
# sqlite3, json, base64, hashlib, hmac, time, datetime, os, sys, re, urllib
# asyncio, threading, pathlib, uuid, random, secrets, traceback, logging
# collections, itertools, functools, copy, pickle, gzip, zipfile, shutil
# tempfile, io, csv, xml, html, http, socket, ssl, subprocess, signal
# inspect, ast, enum, math, decimal, array, queue, contextlib, warnings
# typing, dataclasses, weakref, gc, platform, stat, glob, fnmatch, mimetypes
# email, smtplib, imaplib, poplib, ftplib, telnetlib, configparser, argparse